+++ llarm-emu/src/cpu/instructions/thumb/math.cpp
@@ -132,7 +132,7 @@ void INSTRUCTIONS::thumb::math::ADD5(const u16 code) {
 
     const id::reg Rd_id = reg.thumb_fetch_reg_id(code, 8, 10);
 
-    reg.write(Rd_id, (reg.read(id::reg::PC) & 0xFFFFFFFC) + (immed_8 << 2));
+    reg.write(Rd_id, (reg.read(id::reg::PC) & 0xFFFFFFFC) + u32(immed_8 << 2));
 }
 
 
@@ -143,7 +143,7 @@ void INSTRUCTIONS::thumb::math::ADD6(const u16 code) {
     const u8 immed_8 = llarm::util::bit_range<u8>(code, 0, 7);
     const id::reg Rd_id = reg.thumb_fetch_reg_id(code, 8, 10);
 
-    reg.write(Rd_id, (reg.read(id::reg::SP) + (immed_8 << 2)));
+    reg.write(Rd_id, (reg.read(id::reg::SP) + u32(immed_8 << 2)));
 }
 
 
@@ -153,7 +153,7 @@ void INSTRUCTIONS::thumb::math::ADD6(const u16 code) {
 void INSTRUCTIONS::thumb::math::ADD7(const u16 code) {
     const u8 immed_7 = llarm::util::bit_range<u8>(code, 0, 6);
 
-    reg.write(id::reg::SP, (reg.read(id::reg::SP) + (immed_7 << 2)));
+    reg.write(id::reg::SP, (reg.read(id::reg::SP) + u32(immed_7 << 2)));
 }
 
 
@@ -257,7 +257,7 @@ void INSTRUCTIONS::thumb::math::SUB3(const u16 code) {
 void INSTRUCTIONS::thumb::math::SUB4(const u16 code) {
     const u8 immed_7 = llarm::util::bit_range<u8>(code, 0, 6);
 
-    reg.write(id::reg::SP, (reg.read(id::reg::SP) - (immed_7 << 2)));
+    reg.write(id::reg::SP, (reg.read(id::reg::SP) - u32(immed_7 << 2)));
 }
